Add files via upload
The R script in question leverages the Martini package to construct networks from Single Nucleotide Variants that are frequently observed in case-control samples. The script adheres to a specific file format. Later, the Circilize package is employed to generate a Manhattan plot, with certain variants being accentuated using various shapes and colors.